Move document_index_created to its proper location.
