+++ api/events/event_handlers/create_document_index.py
@@ -5,7 +5,7 @@ import click
 from werkzeug.exceptions import NotFound
 
 from core.indexing_runner import DocumentIsPausedError, IndexingRunner
-from events.event_handlers.document_index_event import document_index_created
+from api.events.document_index_event import document_index_created
 from extensions.ext_database import db
 from libs.datetime_utils import naive_utc_now
 from models.dataset import Document
